Make constructor receiving inner exception public (#33)
Some mocking libraries (i.e. JustMock) throw unit testing framework specific exceptions so that 
the mock verification exceptions integrate better with the test framework. 

By making this constructor public, we make it easier for them to integrate by allowing direct 
creation passing the inner mocking library exception in the constructor. Note that this constructor 
is public in the `Exception` base class, so it's not unusual for derived classes to also expose it 
publicly.